inputable map size, some tweaks, debug info for island generator

+++ src/hexmap.rs
@@ -2,15 +2,15 @@ use hex::{RATIO, Hex};
 
 #[derive(Deserialize, Debug, Clone)]
 pub struct HexMap {
-    pub size_x: i32,
-    pub size_y: i32,
+    pub size_x: u32,
+    pub size_y: u32,
     pub field: Vec<Hex>,
     pub absolute_size_x: f32,
     pub absolute_size_y: f32,
 }
 
 impl HexMap {
-    pub fn new(size_x: i32, size_y: i32) -> HexMap {
+    pub fn new(size_x: u32, size_y: u32) -> HexMap {
         let mut field: Vec<Hex> = Vec::with_capacity((size_x * size_y) as usize);
         for i in 0..(size_x * size_y) {
             let coords = HexMap::index_to_coords(i, size_x, size_y);
@@ -23,8 +23,8 @@ impl HexMap {
         HexMap{size_x, size_y, field, absolute_size_x, absolute_size_y}
     }
 
-    pub fn coords_to_index(x: i32, y: i32, size_x: i32, size_y: i32) -> usize {
-        let base = y * size_x;
+    pub fn coords_to_index(x: i32, y: i32, size_x: u32, size_y: u32) -> usize {
+        let base = y * size_x as i32;
         let offset = y / 2;
         let index = (base + x + offset) as usize;
         if index > (size_x * size_y) as usize {
@@ -33,12 +33,12 @@ impl HexMap {
         index
     }
 
-    pub fn index_to_coords(i: i32, size_x: i32, size_y: i32) -> (i32, i32) {
+    pub fn index_to_coords(i: u32, size_x: u32, size_y: u32) -> (i32, i32) {
         if i >= size_x * size_y {
             panic!{"index {} out of range", i};
         }
-        let line = i / size_x;
-        let pos = i - line * size_x - (line / 2);
+        let line = i as i32 / size_x as i32;
+        let pos = i as i32 - line * size_x as i32 - (line / 2);
         (pos, line)
     }
 }
